requirements.txt
Flask
flask-swagger
Flask-Migrate
sqlalchemy
python-dotenv